requirements.txt
matrix-nio[e2e]==0.24.0
asyncpg==0.29.0
httpx==0.27.0
python-dotenv==1.0.1
